requirements.txt
transformers
sentence-transformers
torch
pandas
PyPDF2
python-dotenv